refactor(gitguard): refactor commit command to use new commit coordinator
• Refactored commit command to utilize new commit coordinator
• Removed legacy commit logic and related imports
• Simplified commit execution flow by delegating to coordinator
• Improved maintainability and modularity of commit handling
• Removed redundant functions and services
• No functional changes expected
• Ensured backward compatibility with legacy commit analysis export
• May require regression testing to confirm no breaking changes
