+++ packages/gitguard/e2e/test/large-commits.test.ts
@@ -54,9 +54,14 @@ const scenarios: TestScenario[] = [
 export const largeCommitsTest: E2ETest = {
   name: "Large Commits Detection",
   scenarios,
-  async run(logger: LoggerService): Promise<TestResult[]> {
+  async run(
+    logger: LoggerService,
+    selectedScenarios?: TestScenario[],
+  ): Promise<TestResult[]> {
     const results: TestResult[] = [];
-    for (const scenario of scenarios) {
+    const scenariosToRun = selectedScenarios || scenarios;
+
+    for (const scenario of scenariosToRun) {
       results.push(await runScenario(scenario, logger));
     }
     return results;
